// risk_rpg/src/app/public/models/enums/colorEnum.ts
export enum ColorEnum {
  WHITE = 0,
  BLACK = 1,
  GRAY = 2,
  RED = 3,
  GREEN = 4,
  BLUE = 5,
  YELLOW = 6,
  MAGENTA = 7,
  CYAN = 8,
  ORANGE = 9,
  VIOLET = 10,
  BROWN = 11,
  TURQUOISE = 12,
  OLIVE = 13,
  NAVY = 14,
  PURPLE = 15,

  LIGHT_GRAY = 16,
  DARK_GRAY = 17,
  LIGHT_RED = 18,
  DARK_RED = 19,
  LIGHT_GREEN = 20,
  DARK_GREEN = 21,
  LIGHT_BLUE = 22,
  DARK_BLUE = 23,
}